refactor: scroll-linked 컴포넌트 리팩토링 작업
- 필요없는 코드 제거
- 커스텀 훅 분리작업
